split out the schema-building functions
